ws implementation base
