+++ package.json
@@ -1,0 +1,20 @@
+{
+  "name": "aiosonic",
+  "version": "1.0.0",
+  "description": "![github status](https://github.com/sonic182/aiosonic/actions/workflows/python.yml/badge.svg) [![Coverage Status](https://coveralls.io/repos/github/sonic182/aiosonic/badge.svg?branch=master)](https://coveralls.io/github/sonic182/aiosonic?branch=master) [![PyPI version](https://badge.fury.io/py/aiosonic.svg)](https://badge.fury.io/py/aiosonic) [![Documentation Status](https://readthedocs.org/projects/aiosonic/badge/?version=latest)](https://aiosonic.readthedocs.io/en/latest/?badge=latest) [![Discord](https://img.shields.io/discord/898929656969965648)](https://discord.gg/e7tBnYSRjj)",
+  "main": "index.js",
+  "directories": {
+    "test": "tests"
+  },
+  "scripts": {
+    "ws-server": "node tests/nodeapps/ws_server.mjs"
+  },
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "ws": "^8.18.0"
+  },
+  "devDependencies": {
+    "typescript": "^5.7.3"
+  }
+}
